+++ .package
@@ -1,2 +1,1 @@
-file logger.c
 file logger.h
